ACTIVITY-REPORT-SILENT-7D — silent-check uses 7d window, matches label
The silent-bot check label says 'enabled but 0 fires in 7d' but
_print_silent_check was being called with the SHORT (1d) window —
flagged bot_cs2_value_v1 as silent every day despite having 3 fires
in the trailing 7d. One-line: pass long instead of short.

+++ scripts/esports/cs2_bot_activity_report.py
@@ -208,7 +208,9 @@ def main() -> int:
 
     _print_bot_table(short, args.days)
     _print_bot_table(long, args.long_days)
-    _print_silent_check(short)
+    _print_silent_check(long)   # silent-check uses the LONGER window so
+                                # low-volume bots aren't falsely flagged on
+                                # a quiet 1d slice (label says "0 fires in 7d").
     _print_funnel(funnel, args.days)
     _print_recent(recent, args.recent)
